Add basic environment variables to all generated scripts
The vars BUILD_DIR and PROJECT_ROOT should be available in all scripts so that scripts can make use of these variables and still be run independently of the main build script (which used to be the place to set these).

More variables might be added later (especially VERSION, but it's not really used anywhere other than the AppImage build script).

Note: the creation of the shell script generator should be handled differently, it's a lot of code duplication and passing variables around right now.